FE: issue 022 get current location fix
