+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React from 'react';
 import './App.css';
 import {HashRouter, Route, Routes} from "react-router-dom";
 import MainPage from "./pages/MainPage";
@@ -8,35 +8,11 @@ import SignInPage from "./pages/SignInPage";
 import SignUpPage from "./pages/SignUpPage";
 import {ToastContainer} from "react-toastify";
 import RouteDetailsPage from "./pages/RouteDetailsPage";
-import axios from "axios";
-import {LocationReturn} from "./model/LocationReturn";
 
 function App() {
 
     const {setRequest, saveFoundRoutes} = useMyRoutes()
-    const [currentAddress, setCurrentAddress] = useState<LocationReturn>({
-        address: {city:"",
-            country_code:"",
-            postcode:"",
-            road:"",
-            state:"",
-            suburb:""},
-        display_name: "",
-        lat: "",
-        lon: "",
-        osm_id: undefined
-    });
 
-
-
-    function getCurrentLocation(lat:number, lon:number) {
-        console.log(lat, lon)
-        axios.get("https://nominatim.openstreetmap.org/reverse?lat="+lat+"&lon="+lon+"&format=json")
-            .then((response)=> {return response.data })
-            .then((data)=>{setCurrentAddress(data)})
-            .finally(()=>console.log("places are: ", currentAddress))
-            .catch((err)=>console.log("err: ", err))
-    }
 
 
     return (
@@ -49,8 +25,7 @@ function App() {
                 <Route path = {"/"} element = {<MainPage saveFoundRoutes={saveFoundRoutes} setRequest={setRequest}/>}/>
                 <Route path = {"/sign-in"} element = {<SignInPage/>}/>
                 <Route path = {"/sign-up"} element = {<SignUpPage/>}/>
-                <Route path = {"/routes/:id"} element = {<RoutesPage saveFoundRoutes={saveFoundRoutes} setRequest={setRequest}
-                                                                     getCurrentLocation={getCurrentLocation} currentAddress={currentAddress}/>}/>
+                <Route path = {"/routes/:id"} element = {<RoutesPage saveFoundRoutes={saveFoundRoutes} setRequest={setRequest}/>}/>
                 <Route path = {"/route/:id/details"} element = {<RouteDetailsPage/>}/>
             </Routes>
         </HashRouter>
